Update ex06Sig.c
Line 61: type int is causing Segmentation Fault. Type size_t seems to work fine. 
Line 75: Should we cast slen of type size_t to int in order to mute the warning?

+++ SPSA18/OpenSSL/ex06Sig.c
@@ -58,7 +58,7 @@ main(int argc, char *argv[])
         return 1;
     }
  
-    int slen;
+    size_t slen;
     if(1 != EVP_DigestSignFinal(mdctx,NULL,(size_t *)&slen)){
         fprintf(stderr,"Could not finalize\n");
         return 1;
